--- TourGuide/src/main/java/com/openclassrooms/tourguide/TourGuideModule.java
@@ -1,32 +1,0 @@
-package com.openclassrooms.tourguide;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import gpsUtil.GpsUtil;
-import org.springframework.core.task.TaskExecutor;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-import rewardCentral.RewardCentral;
-import com.openclassrooms.tourguide.service.RewardsService;
-
-@Configuration
-public class TourGuideModule {
-
-	private ThreadPoolTaskExecutor taskExecutor;
-
-	@Bean
-	public GpsUtil getGpsUtil() {
-		return new GpsUtil();
-	}
-	
-	@Bean
-	public RewardsService getRewardsService() {
-		return new RewardsService(getGpsUtil(), getRewardCentral(), taskExecutor);
-	}
-	
-	@Bean
-	public RewardCentral getRewardCentral() {
-		return new RewardCentral();
-	}
-	
-}
