+++ TourGuide/src/main/java/com/openclassrooms/tourguide/config/LibsModuleConfig.java
@@ -1,0 +1,22 @@
+package com.openclassrooms.tourguide.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import gpsUtil.GpsUtil;
+import rewardCentral.RewardCentral;
+
+@Configuration
+public class LibsModuleConfig {
+
+    @Bean
+	public GpsUtil getGpsUtil() {
+		return new GpsUtil();
+	}
+	
+	@Bean
+	public RewardCentral getRewardCentral() {
+		return new RewardCentral();
+	}
+	
+}
